# src/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/authentication/login": {
            "post": {
                "description": "Creates a token for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Creates a token",
                "parameters": [
                    {
                        "description": "User credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LoginUserPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Healthcheck endpoint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "Healthcheck",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetches a user profile by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Fetches a user profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/weather/coords/{latitude}/{longitude}": {
            "get": {
                "description": "Fetches the current weather data for a specified latitude and longitude",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Fetches current weather by coordinates (latitude and longitude)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Latitude",
                        "name": "latitude",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Longitude",
                        "name": "longitude",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.WeatherResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/weather/forecast/coords/{latitude}/{longitude}": {
            "get": {
                "description": "Fetches the weather forecast data for a specified latitude and longitude",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Fetches weather forecast by coordinates (latitude and longitude)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Latitude",
                        "name": "latitude",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Longitude",
                        "name": "longitude",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.ForecastResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/weather/forecast/{city}": {
            "get": {
                "description": "Fetches the weather forecast data for a specified city",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Fetches weather forecast by city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City Name",
                        "name": "city",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.ForecastResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/weather/open-meteo/coords/{latitude}/{longitude}": {
            "get": {
                "description": "Fetches the open meteo data for a specified latitude and longitude",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Fetches open meteo by coordinates (latitude and longitude)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Latitude",
                        "name": "latitude",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Longitude",
                        "name": "longitude",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.OpenMeteoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/weather/{city}": {
            "get": {
                "description": "Fetches the current weather data for a specified city",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Fetches current weather by city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City Name",
                        "name": "city",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.WeatherResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "main.LoginUserPayload": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 3
                }
            }
        },
        "store.City": {
            "type": "object",
            "properties": {
                "coord": {
                    "$ref": "#/definitions/store.Coord"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "population": {
                    "type": "integer"
                },
                "sunrise": {
                    "type": "integer"
                },
                "sunset": {
                    "type": "integer"
                },
                "timezone": {
                    "type": "integer"
                }
            }
        },
        "store.Clouds": {
            "type": "object",
            "properties": {
                "all": {
                    "type": "integer"
                }
            }
        },
        "store.Coord": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                }
            }
        },
        "store.DailyForecast": {
            "type": "object",
            "properties": {
                "apparent_temperature_max": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "apparent_temperature_min": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "time": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "weather_code": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "store.DailyUnits": {
            "type": "object",
            "properties": {
                "apparent_temperature_max": {
                    "type": "string"
                },
                "apparent_temperature_min": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "weather_code": {
                    "type": "string"
                }
            }
        },
        "store.ForecastItem": {
            "type": "object",
            "properties": {
                "clouds": {
                    "$ref": "#/definitions/store.Clouds"
                },
                "dt": {
                    "type": "integer"
                },
                "dt_txt": {
                    "type": "string"
                },
                "main": {
                    "$ref": "#/definitions/store.Main"
                },
                "pop": {
                    "type": "number"
                },
                "sys": {
                    "$ref": "#/definitions/store.Sys"
                },
                "visibility": {
                    "type": "integer"
                },
                "weather": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.Weather"
                    }
                },
                "wind": {
                    "$ref": "#/definitions/store.Wind"
                }
            }
        },
        "store.ForecastResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "$ref": "#/definitions/store.City"
                },
                "cnt": {
                    "type": "integer"
                },
                "cod": {
                    "type": "string"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.ForecastItem"
                    }
                },
                "message": {
                    "type": "integer"
                }
            }
        },
        "store.HourlyForecast": {
            "type": "object",
            "properties": {
                "temperature_2m": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "time": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "store.HourlyUnits": {
            "type": "object",
            "properties": {
                "temperature_2m": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "store.Main": {
            "type": "object",
            "properties": {
                "feels_like": {
                    "type": "number"
                },
                "grnd_level": {
                    "type": "integer"
                },
                "humidity": {
                    "type": "integer"
                },
                "pressure": {
                    "type": "integer"
                },
                "sea_level": {
                    "type": "integer"
                },
                "temp": {
                    "type": "number"
                },
                "temp_kf": {
                    "type": "number"
                },
                "temp_max": {
                    "type": "number"
                },
                "temp_min": {
                    "type": "number"
                }
            }
        },
        "store.OpenMeteoResponse": {
            "type": "object",
            "properties": {
                "daily": {
                    "$ref": "#/definitions/store.DailyForecast"
                },
                "daily_units": {
                    "$ref": "#/definitions/store.DailyUnits"
                },
                "elevation": {
                    "type": "number"
                },
                "generationtime_ms": {
                    "type": "number"
                },
                "hourly": {
                    "$ref": "#/definitions/store.HourlyForecast"
                },
                "hourly_units": {
                    "$ref": "#/definitions/store.HourlyUnits"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "timezone": {
                    "type": "string"
                },
                "timezone_abbreviation": {
                    "type": "string"
                },
                "utc_offset_seconds": {
                    "type": "integer"
                }
            }
        },
        "store.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "store.Sys": {
            "type": "object",
            "properties": {
                "pod": {
                    "type": "string"
                }
            }
        },
        "store.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/store.Role"
                },
                "role_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "store.Weather": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "main": {
                    "type": "string"
                }
            }
        },
        "store.WeatherResponse": {
            "type": "object",
            "properties": {
                "clouds": {
                    "$ref": "#/definitions/store.Clouds"
                },
                "cod": {
                    "type": "integer"
                },
                "coord": {
                    "$ref": "#/definitions/store.Coord"
                },
                "dt": {
                    "description": "Timestamp",
                    "type": "integer"
                },
                "main": {
                    "type": "object",
                    "properties": {
                        "feels_like": {
                            "type": "number"
                        },
                        "grnd_level": {
                            "type": "integer"
                        },
                        "humidity": {
                            "type": "integer"
                        },
                        "pressure": {
                            "type": "integer"
                        },
                        "sea_level": {
                            "type": "integer"
                        },
                        "temp": {
                            "type": "number"
                        },
                        "temp_max": {
                            "type": "number"
                        },
                        "temp_min": {
                            "type": "number"
                        }
                    }
                },
                "name": {
                    "type": "string"
                },
                "sys": {
                    "type": "object",
                    "properties": {
                        "country": {
                            "type": "string"
                        },
                        "sunrise": {
                            "type": "integer"
                        },
                        "sunset": {
                            "type": "integer"
                        }
                    }
                },
                "timezone": {
                    "type": "integer"
                },
                "visibility": {
                    "type": "integer"
                },
                "weather": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.Weather"
                    }
                },
                "wind": {
                    "$ref": "#/definitions/store.Wind"
                }
            }
        },
        "store.Wind": {
            "type": "object",
            "properties": {
                "deg": {
                    "type": "integer"
                },
                "gust": {
                    "type": "number"
                },
                "speed": {
                    "type": "number"
                }
            }
        }
    }
}